+++ website/contents/HealthSection.content.ts
@@ -57,7 +57,7 @@ const pageContent = {
     },
     paragraph1: t({
       en: "",
-      tr: "Sağlıklı bir yaşam, sadece güçlü bir bedene sahip olmak değildir; aynı zamanda, kişinin bütün potansiyelini geliştiren bir zihnin varlığına bağlıdır. Zihinsel sağlık, odaklanma, fırsatlar, yetenekler ve diğer şans unsurları aracılığıyla büyümeyi besler — bunların tümü de doğanın adalet terazisine bağlıdır. Çevremizle olan ilişkilerin getirileri olan doğal yeteneklere sahip olabiliriz.\n\nDoğanın bütünlüğünde, bir bebeğin gelişimi esnasında oluşan hastalıkların ebeveylerin çevreleriyle olan ilişkilerine bağlı olduğunu anlayabiliriz. Kalıtsal hastalıkların iyileşmesi için daha fazla çaba göstermek gerekir. Muhtemelen çoğu genetik hastalığın tedavisi sadece bir sonraki nesilde gerçekleşebilir. Çünkü bizden önceki ve mevcut nesiller aracılığıyla doğaya verilen zararın incelenmesi ve telafi edilmesi gerekir. İnsanı doğanın bir parçası olarak görürken, diğer canlıları unutmamalıyız. Bu yüzden, bazı geleneklerin tekrar yorumlanması gerekebilir: Öz eleştiri ve farkındalık her iyileşmenin başlangıcıdır.\n\n**Herhangi mikrop, virüs veya kimyasal vücudumuza girdiğinde, seçtiğimiz eylemlere göre vücudumuzda yön bulur.** Zararlı davranış kalıplarında ısrar etmek, bunların vücudumuza zarar vererek yayılmasına yol açar. Çevremizle birlikte kendi vücut sistemimizi anladıkça ve eylemlerimizi farkındalık, iyileşme ve yenilenme ile yönlendirdikçe, vücudumuz dengesini yeniden kazanır ve zararlı maddelerden kurtulmanın bir yolunu bulur. Doğayla uyumumuzun çok iyi olduğu durumlarda, vücudumuza verilen bilinçli zararlar sorumlularının doğada cezalandırılmasına vesile olur.\n\n**Organ-Eylem-Sonuç Benzetmesi yaşadığımız tüm kazalar için de geçerlidir.** Herhangi bir kaza sonucunda vücudumuzda oluşan hasarlar, kazadan önceki eylemlerimizin bir yansıması olarak bize bir ders vermeyi amaçlar. Yaşadığımız bir kazanın oluşma şekli, sebep olan eylemimizi açıklar. Zihnimizdeki odak noktamızı ve vücudumuzu nasıl kullandığımızı açıklar.\n\nDoğadaki dengenin sağlanabilmesi için, Yaradan aracılığıyla hafızamız, düşüncelerimiz ve hislerimiz şekil değiştirebilir, sınanabilir ve silinebilir. Doğayla kurulan dengeli bir bağ, sağlam bir zihinsel yapıya imkan tanır. Hatalarımızı telafi etme fırsatı yeniden doğsa bile, bazı hatalar geri dönülemeyecek sonuçlara vesile olur.",
+      tr: "Sağlıklı bir yaşam, sadece güçlü bir bedene sahip olmak değildir; aynı zamanda, kişinin bütün potansiyelini geliştiren bir zihnin varlığına bağlıdır. Zihinsel sağlık, odaklanma, fırsatlar, yetenekler ve diğer şans unsurları aracılığıyla büyümeyi besler — bunların tümü de doğanın adalet terazisine bağlıdır. Çevremizle olan ilişkilerin getirileri olan doğal yeteneklere sahip olabiliriz.\n\nDoğanın bütünlüğünde, bir bebeğin gelişimi esnasında oluşan hastalıkların ebeveylerin çevreleriyle olan ilişkilerine bağlı olduğunu anlayabiliriz. Kalıtsal hastalıkların iyileşmesi için daha fazla çaba göstermek gerekir. Muhtemelen çoğu genetik hastalığın tedavisi sadece bir sonraki nesilde gerçekleşebilir. Çünkü bizden önceki ve mevcut nesiller aracılığıyla doğaya verilen zararın incelenmesi ve telafi edilmesi gerekir. İnsanı doğanın bir parçası olarak görürken, diğer canlıları unutmamalıyız. Bu yüzden, bazı geleneklerin tekrar yorumlanması gerekebilir: Öz eleştiri ve farkındalık her iyileşmenin başlangıcıdır.\n\n**Herhangi mikrop, virüs veya kimyasal vücudumuza girdiğinde, seçtiğimiz eylemlere göre vücudumuzda yön bulur.** Zararlı davranış kalıplarında ısrar etmek, bunların vücudumuza zarar vererek yayılmasına yol açar. Çevremizle birlikte kendi vücut sistemimizi anladıkça ve eylemlerimizi farkındalık, iyileşme ve yenilenme ile yönlendirdikçe, vücudumuz dengesini yeniden kazanır ve zararlı maddelerden kurtulmanın bir yolunu bulur. Doğayla uyumumuzun çok iyi olduğu durumlarda, vücudumuza verilen bilinçli zararlar sorumlularının doğada cezalandırılmasına vesile olur.\n\n**Organ-Eylem-Sonuç Benzetmesi yaşadığımız tüm kazalar için de geçerlidir.** Herhangi bir kaza sonucunda vücudumuzda oluşan hasarlar, kazadan önceki eylemlerimizin bir yansıması olarak bize bir ders vermeyi amaçlar. Yaşadığımız bir kazanın oluşma şekli, sebep olan eylemimizi açıklar. Zihnimizdeki odak noktamızı ve vücudumuzu nasıl kullandığımızı açıklar.\n\nDoğadaki dengenin sağlanabilmesi için, Yaradan aracılığıyla hafızamız, düşüncelerimiz ve hislerimiz şekil değiştirebilir, sınanabilir ve silinebilir. Doğayla kurulan dengeli bir bağ, sağlam bir zihinsel yapıya imkan tanır. Hatalarımızı telafi etme fırsatı yeniden doğsa bile, bazı hatalar geri dönülemeyecek sonuçlara vesile olur.\n\n*Doğal afetlerin vücudumuzdakine benzer bir sebep-sonuç ilişkisiyle gerçekleştiklerini gözlemleyebilmemiz için kendimizi ve çevremizi daha iyi tanımamız gerekir.*",
       de: "",
       fr: ""
     }),
