+++ website/contents/MovementSection.content.ts
@@ -14,7 +14,65 @@ const pageContent = {
       tr: "Mantıksal Maneviyat, bireysel sağlığımızın tüm çevremizle olan ilişkilere bağlı olduğunu kanıtlıyor. Bireysel hataların cezasının, aynı zamanda tüm toplum ve çevrede olumsuz şekilde görülebildiğini açığa çıkarıyor. Mevcut sağlık çözümleri hastalıklara kalıcı bir çözüm bulmak yerine, sadece sonuçlara yama uyguluyor. Herhangi bir tedavi veya ilaç kullanımından sonra yaşamına geri dönmek zorunda olan kişilerin yaşam koşullarında iyileşme olmadığından dolayı doğada hastalıklar tekrar ediyor veya şekil değiştiriyor.\n\nDoğanın kanunları yaşamdaki eşitliği ve özgürlüğü hissedebilir bir biçimde uygularken, güncel hukuk düzenleri doğa için gerekli koşulları sağlayamıyor. Günümüzde en modern hukuk sistemi bile insan haklarına aykırı bir biçimde uygulanmış oluyor. Eylemleri gerçekleştiren kişilerin doğa kanunlarıyla yüzleşmeleri ve olumsuzlukları yaşıyor olmaları, mevcut yönetim biçimlerinin tüm canlıların eşitliğini sağlamasına yönelik değişim ihtiyacını işaret ediyor.\n\nTüm canlıların sağlıklı bir toplumda hiç hastalanmadan yaşayabilecekleri bir düzen kurmak mümkün. Doğanın ilkelerini derinlemesine anlamak ve uygulamalarını güncel yönetim sistemlerine entegre edebilmek için bu projeyi başlatma ihtiyacı duydum.",
       de: "",
       fr: "",
-    })
+    }),
+    items: [
+      {
+        title: t({
+          en: "Environment",
+          tr: "Çevre",
+          de: "Umwelt",
+          fr: "Environnement"
+        }),
+        subtitle: t({
+          en: "",
+          tr: "",
+          de: "",
+          fr: ""
+        }),
+      },
+      {
+        title: t({
+          en: "",
+          tr: "Sağlık",
+          de: "",
+          fr: ""
+        }),
+        subtitle: t({
+          en: "",
+          tr: "",
+          de: "",
+          fr: ""
+        }),
+      },
+      {
+        title: t({
+          en: "",
+          tr: "Sağlık",
+          de: "",
+          fr: ""
+        }),
+        subtitle: t({
+          en: "",
+          tr: "",
+          de: "",
+          fr: ""
+        }),
+      },
+      {
+        title: t({
+          en: "",
+          tr: "Sağlık",
+          de: "",
+          fr: ""
+        }),
+        subtitle: t({
+          en: "",
+          tr: "",
+          de: "",
+          fr: ""
+        }),
+      }
+    ],
   },
 } satisfies Dictionary;
